requirements.txt
cryptography
Django
djangorestframework
pem
PyJWT
setuptools
